sweep: #7943 Some refactoring
